% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp.r
\name{evaluate.gp}
\alias{evaluate.gp}
\title{Evaluate a Gaussian process density function}
\usage{
evaluate.gp(fit, x)
}
\arguments{
\item{fit}{mvd.density object}

\item{x}{Matrix or vector of samples. For matrices, rows are samples and columns are variables.}
}
\description{
description
}
