% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vine_copula.r
\name{evaluate.vine.copula}
\alias{evaluate.vine.copula}
\title{Evaluate a vine copula}
\usage{
evaluate.vine.copula(fit, x, log = F)
}
\arguments{
\item{fit}{An mvd.density object obtained from fit.vine.copula}

\item{x}{Matrix or vector of samples at which to evaluate the vine copula. For matrices, rows are samples and columns are variables.}

\item{log}{Return log probability density}
}
\description{
description
}
