% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmm.r
\name{fit.gmm}
\alias{fit.gmm}
\title{Fit a Gaussian mixture with a specific number of components.}
\usage{
fit.gmm(x, K, epsilon = 1e-05, maxsteps = 1000, verbose = F)
}
\arguments{
\item{x}{Matrix or vector of samples. For matrices, rows are samples and columns are variables.}

\item{K}{Integer specifying the number of components of the Gaussian mixture.}

\item{epsilon}{For the EM algorithm, stop when the difference in log likelihood is less than this epsilon.}

\item{maxsteps}{Maximum number of steps to take in the EM algorithm. When the maximum number is reached, the current fit will be returned and a warning will be issued.}

\item{verbose}{Display the fitting progress by showing the likelihood at every iteration.}
}
\description{
description
}
