% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp.r
\name{fit.gp}
\alias{fit.gp}
\title{Fit a Gaussian process density function.}
\usage{
fit.gp(x, p, kernel, l = 1, optimize = T, normalize = T, sigman = 1e-12,
  verbose = F)
}
\arguments{
\item{x}{Matrix or vector of samples. For matrices, rows are samples and columns are variables.}

\item{p}{Probability density at the sample locations; does not have to be normalized. Note that the probabilities should not be log transformed.}

\item{kernel}{Which kernel to use, can be "se" or "matern32".}

\item{l}{If optimize == TRUE, this should be a vector of size two with the lower and upper bound for the length scale; if optimize == FALSE, this is the fixed length scale.}

\item{optimize}{Boolean specifying whether l (and possibly sigman) should be optimized.}

\item{normalize}{Boolean specifying whether to normalize the GP such that it integrates to 1.}

\item{sigman}{Can be either a fixed sigman or, if optimize == TRUE, this can also be a vector of size two specifying the lower and upper bound, and in this case sigman will be optimized along with l.}

\item{verbose}{Display progress during the optimization of l.}
}
\description{
Fit a Gaussian process through the posterior density samples. The GP can optionally be normalized such that the predictive function integrates to 1.
The GP assumes a zero mean, and can use either a squared exponential or Matern-type covariance function. A single length scale parameter is used, which is
optimized by minimizing the RMSE in 5-fold cross validation.
}
