% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde.r
\name{fit.kde}
\alias{fit.kde}
\title{Fit a kernel density estimate}
\usage{
fit.kde(x, adjust = 1, bw.fn = bw.SJ, diagonal = T, verbose = F, ...)
}
\arguments{
\item{x}{Matrix or vector of samples. For matrices, rows are samples and columns are variables.}

\item{adjust}{Scalar multiplication of the bandwidth}

\item{bw.fn}{Function used to calculate the bandwidth. If diagonal = T, bw.fn should accept a vector of 1-dimensional values. If diagonal = F, bw.fn should accept a matrix of samples.}

\item{...}{Further arguments are passed to bw.fn}
}
\description{
description
}
\examples{
x <- rmvnorm(50, c(0, 0), rbind(c(1, 0.5), c(0.5, 1)))
fit <- fit.kde(x)
fit <- fit.kde(x, bw.fn = Hpi, diagonal = F)
}
