% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde.r
\name{fit.kde.transformed}
\alias{fit.kde.transformed}
\title{Fit a kernel density estimate after transforming the variables to an unbounded domain.}
\usage{
fit.kde.transformed(x, bounds, adjust = 1, bw.fn = bw.SJ, diagonal = TRUE,
  verbose = FALSE)
}
\arguments{
\item{x}{Matrix or vector of samples. For matrices, rows are samples and columns are variables.}

\item{bounds}{Dx2 matrix specifying the lower and upper bound for each variable.}

\item{adjust}{Scalar multiplication of the bandwidth}

\item{bw.fn}{Function used to calculate the bandwidth. If diagonal = T, bw.fn should accept a vector of 1-dimensional values. If diagonal = F, bw.fn should accept a matrix of samples.}

\item{diagonal}{If true, estimate a diagonal bandwidth matrix; otherwise estimate a full bandwidth matrix.}

\item{verbose}{No effect.}
}
\description{
The following transformations are used
[0,inf] -> log(x)
[-inf,0] -> log(-x)
[0,1] -> logit(x) = log(x / (1 - x))
[a,b] -> scaled logit(x) = log((a - x) / (x - b))
}
\examples{
x <- exp(mvtnorm::rmvnorm(50, c(0, 0), rbind(c(1, 0.5), c(0.5, 1))))
fit <- fit.kde(x, rbind(c(0, -Inf), c(0, Inf)))
fit <- fit.kde(x, rbind(c(0, -Inf), c(0, Inf)), bw.fn = ks::Hpi, diagonal = FALSE)
}
