% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginals.r
\name{fit.marginal.ecdf.pareto}
\alias{fit.marginal.ecdf.pareto}
\title{Marginal function for use in fit.vine.copula: ecdf + pareto tail}
\usage{
fit.marginal.ecdf.pareto(x, bounds = cbind(rep(-Inf, ncol(x)), rep(Inf,
  ncol(x))), reflect_bounds = T, pareto_threshold = 0.1,
  ecdf_bounds_threshold = 0.001)
}
\arguments{
\item{x}{Matrix or vector of samples. For matrices, rows are samples and columns are variables.}

\item{bounds}{Dx2 matrix specifying the lower and upper bound for each variable.}

\item{reflect_bounds}{If true, reflect samples across bounds to improve the estimate near the boundaries.}

\item{pareto_threshold}{Cumulative probability threshold for the start of the pareto tail}

\item{ecdf_bounds_threshold}{Use a pareto tail only if the density of the kernel density estimate at the boundary is below this value.}
}
\description{
Marginal function for use in fit.vine.copula: ecdf + pareto tail
}
