% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginals.r
\name{fit.marginal.ecdf}
\alias{fit.marginal.ecdf}
\title{Marginal function for use in fit.vine.copula: empirical distribution function}
\usage{
fit.marginal.ecdf(x, bounds = cbind(rep(-Inf, ncol(x)), rep(Inf,
  ncol(x))), reflect_bounds = T)
}
\arguments{
\item{x}{Matrix or vector of samples. For matrices, rows are samples and columns are variables.}

\item{bounds}{Dx2 matrix specifying the lower and upper bound for each variable.}

\item{reflect_bounds}{If true, reflect samples across bounds to improve the estimate near the boundaries.}
}
\description{
Marginal function for use in fit.vine.copula: empirical distribution function
}
