% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginals.r
\name{fit.marginal.parametric}
\alias{fit.marginal.parametric}
\title{Marginal function for use in fit.vine.copula: single parametric distribution}
\usage{
fit.marginal.parametric(x, bounds = cbind(rep(-Inf, ncol(x)), rep(Inf,
  ncol(x))))
}
\arguments{
\item{x}{Matrix or vector of samples. For matrices, rows are samples and columns are variables.}

\item{bounds}{Dx2 matrix specifying the lower and upper bound for each variable.}
}
\description{
Marginal function for use in fit.vine.copula: single parametric distribution
}
