% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vine_copula.r
\name{fit.vine.copula}
\alias{fit.vine.copula}
\title{Fit a vine copula}
\usage{
fit.vine.copula(x, marginalfn, bounds = cbind(rep(-Inf, ncol(x)), rep(Inf,
  ncol(x))), trunclevel = NA, indeptest = F, verbose = F)
}
\arguments{
\item{x}{Matrix or vector of samples. For matrices, rows are samples and columns are variables.}

\item{marginalfn}{One of the marginal fitting functions.}

\item{bounds}{Dx2 matrix of boundaries}

\item{trunclevel}{Relayed to the VineCopula::RVineStructureSelect}

\item{indeptest}{Relayed to VineCopula::RVineStructureSelect}

\item{verbose}{Relayed to VineCopula::RVineStructureSelect}
}
\description{
Fit a vine copula density function, using one of the marginal function for the marginal transformations, and using VineCopula::RVineStructureSelect for the vine copula function.
}
