% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmm.r
\name{gmm.BIC}
\alias{gmm.BIC}
\title{Calculate BIC of a Gaussian mixture across a range of number of components.}
\usage{
gmm.BIC(x, K = 1:6, optimal.only = F, epsilon = 1e-05, maxsteps = 1000,
  verbose = F)
}
\arguments{
\item{x}{Matrix or vector of samples. For matrices, rows are samples and columns are variables.}

\item{K}{Vector specifying the number of components to test}

\item{optimal.only}{If TRUE, directly return only the GMM with optimal number of components; otherwise return a structure with fits for all tested number of components}

\item{epsilon}{See fit.gmm}

\item{maxsteps}{See fit.gmm}

\item{verbose}{Display the fitting progress.}
}
\description{
description
}
