% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmm.r
\name{gmm.truncated.BIC}
\alias{gmm.truncated.BIC}
\title{Calculate BIC of a truncated Gaussian mixture across a range of number of components}
\usage{
gmm.truncated.BIC(x, K = 1:6, bounds = cbind(rep(-Inf, ncol(x)),
  rep(Inf, ncol(x))), optimal.only = F, min_cov = NULL,
  epsilon = 1e-05, maxsteps = 1000, verbose = F)
}
\arguments{
\item{x}{Matrix or vector of samples. For matrices, rows are samples and columns are variables.}

\item{K}{Vector specifying the number of components to test}

\item{bounds}{Dx2 matrix specifying the lower and upper bound for each variable.}

\item{optimal.only}{If TRUE, directly return only the GMM with optimal number of components; otherwise return a structure with fits for all tested number of components}

\item{epsilon}{See fit.gmm.truncated}

\item{maxsteps}{See fit.gmm.truncated}

\item{verbose}{Display the fitting progress.}
}
\description{
description
}
