% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdf.r
\name{mvd.cdf}
\alias{mvd.cdf}
\title{Evaluate the cumulative distribution function of a density approximation}
\usage{
mvd.cdf(fit, x, log = FALSE)
}
\arguments{
\item{fit}{An mvd.density object obtained from one of the density fitting functions.}

\item{x}{Matrix or vector of positions at which to evaluate the density function.}

\item{log}{Boolean which specifies whether to return the cumulative probability in log space.}
}
\description{
Currently only implemented for Gaussian mixtures.
}
