% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.r
\name{mvd.correct_p_for_transformation}
\alias{mvd.correct_p_for_transformation}
\title{Probability density correction for variable transformation}
\usage{
mvd.correct_p_for_transformation(x, bounds, p, log = T)
}
\arguments{
\item{x}{Matrix or vector of samples. For matrices, rows are samples and columns are variables.}

\item{bounds}{Dx2 matrix specifying the lower and upper bound for each variable.}

\item{p}{Probability densities to correct}

\item{log}{Specifies whether p is in log scale}
}
\description{
Probability density correction for variable transformation
}
