% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_likelihood.r
\name{mvd.marginal.likelihood}
\alias{mvd.marginal.likelihood}
\title{Calculate a marginal likelihood estimate from a density fit}
\usage{
mvd.marginal.likelihood(fit, x, p, log)
}
\arguments{
\item{fit}{An mvd.density object obtained from one of the density fitting functions.}

\item{x}{Matrix or vector of positions at which to evaluate the density function.}

\item{p}{Unnormalized posterior probability values.}

\item{log}{Boolean which specifies whether p is in log scale. If true, the marginal likelihood is also return in log scale.}
}
\description{
Estimates the marginal likelihood from a set of Monte Carlo samples using a fitted posterior density approximation.
For Gaussian process fits, the marginal likelihood estimation only makes sense if the GP was normalized during the fitting.
}
