% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginals.r
\name{mvd.marginal.pdf}
\alias{mvd.marginal.pdf}
\title{Probability density function of a marginal distribution fitted with one of the marginal distribution functions.}
\usage{
mvd.marginal.pdf(marginal, x, log = T)
}
\arguments{
\item{marginal}{mvd.marginal object obtained from one of the marginal fitting functions.}

\item{x}{Matrix or vector of samples. For matrices, rows are samples and columns are variables.}

\item{log}{Whether to return the density in log}
}
\description{
Probability density function of a marginal distribution fitted with one of the marginal distribution functions.
}
