% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginals.r
\name{mvd.marginal.transform}
\alias{mvd.marginal.transform}
\title{Transform variables to U[0,1] using a fitted marginal function}
\usage{
mvd.marginal.transform(x, marginal)
}
\arguments{
\item{x}{Matrix or vector of samples. For matrices, rows are samples and columns are variables.}

\item{marginal}{mvd.marginal object obtained from one of the marginal fitting functions.}
}
\description{
Transform variables to U[0,1] using a fitted marginal function
}
