% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.r
\name{mvd.pdf}
\alias{mvd.pdf}
\title{Evaluate the probability density function of a density approximation}
\usage{
mvd.pdf(fit, x, log = FALSE)
}
\arguments{
\item{fit}{An mvd.density object obtained from one of the density fitting functions.}

\item{x}{Matrix or vector of positions at which to evaluate the density function.}

\item{log}{Boolean which specifies whether to return the probability density in log space.}
}
\description{
description
}
