% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.r
\name{mvd.transform_from_unbounded}
\alias{mvd.transform_from_unbounded}
\title{Reverse transform from an unbounded domain}
\usage{
mvd.transform_from_unbounded(transformed_x, bounds)
}
\arguments{
\item{transformed_x}{Matrix or vector of transformed samples. For matrices, rows are samples and columns are variables.}

\item{bounds}{Dx2 matrix specifying the lower and upper bound for each variable.}
}
\description{
description
}
