% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginals.r
\name{reverse.transform.marginals}
\alias{reverse.transform.marginals}
\title{Reverse transform variables from U[0,1]}
\usage{
reverse.transform.marginals(transformed, marginal)
}
\arguments{
\item{transformed}{Matrix or vector of transformed samples. For matrices, rows are samples and columns are variables.}

\item{marginal}{mvd.marginal object obtained from one of the marginal fitting functions.}
}
\description{
Reverse transform variables from U[0,1]
}
